@@ -2,21 +2,23 @@
 name = "nf-core-stats"
 version = "0.1.0"
 description = "DLT pipeline for nf-core stats collection"
-authors = [
-    { name = "nf-core team" }
-]
+authors = [{ name = "nf-core team" }]
 dependencies = [
-    "dlt[cli,duckdb,motherduck]>=1.4.1",
-    "python-dotenv==1.0.0",
+    "dlt[cli,duckdb,motherduck]>=1.17.1",
+    "python-dotenv==1.1.1",
     "requests>=2.31.0",
     "slack-sdk>=3.33.5",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 
 [project.scripts]
 slack = "python slack_pipeline.py"
 github = "python github_pipeline.py"
 twitter = "python twitter_pipeline.py"
-# TODO
-# lint = "ruff check ."
-# format = "ruff format ."
+
+[tool.ruff]
+line-length = 120
+target-version = "py310"
+
+[tool.ruff.lint]
+select = ["I", "E1", "E4", "E7", "E9", "F", "UP", "N"]
